+++ src/bp/ui-shared/src/TreeView/utils.tsx
@@ -1,0 +1,137 @@
+import _ from 'lodash'
+import { useEffect, useRef } from 'react'
+
+import { DOCUMENT_ICON, FOLDER_ICON } from '.'
+import { BuildTreeParams, ElementRenderer, TreeNode } from './typings'
+
+function splitPath<T>(
+  fullPath: string,
+  nodeData: T,
+  nodeRenderer: ElementRenderer<T>,
+  folderRenderer: ElementRenderer<string>
+) {
+  const splitPath = fullPath.split('/')
+  const nodeLabel = splitPath[splitPath.length - 1]
+  const parentFolders = splitPath.slice(0, splitPath.length - 1)
+
+  const folders: TreeNode<T>[] = []
+  const currentPath: string[] = []
+
+  for (const folder of parentFolders) {
+    currentPath.push(folder)
+    const { label, icon } = folderRenderer(folder)
+
+    folders.push({
+      id: folder,
+      label,
+      icon: icon || FOLDER_ICON,
+      fullPath: currentPath.join('/'),
+      type: 'folder',
+      childNodes: undefined
+    })
+  }
+
+  currentPath.push(nodeLabel)
+
+  const id = currentPath.join('/')
+  const { label, icon } = nodeRenderer(nodeData)
+
+  return {
+    folders,
+    leafNode: { id, label, icon: icon || DOCUMENT_ICON, fullPath: id, type: 'node' }
+  }
+}
+
+export function buildTree<T>({
+  elements,
+  nodeRenderer,
+  filterText,
+  filterProps,
+  folderRenderer = label => ({ label }),
+  postProcessing,
+  pathProps
+}: BuildTreeParams<T>): TreeNode<T>[] {
+  if (!elements) {
+    return []
+  }
+
+  const tree: TreeNode<T> = {
+    id: 'root',
+    label: '<root>',
+    type: 'root',
+    fullPath: '',
+    childNodes: []
+  }
+
+  const lowerCaseFilter = filterText?.toLowerCase()
+
+  for (const nodeData of elements) {
+    const nodePath = nodeData[pathProps || 'path']
+    if (!nodePath) {
+      console.error(`Invalid path`)
+      return []
+    }
+
+    const { folders, leafNode } = splitPath(nodePath, nodeData, nodeRenderer, folderRenderer!)
+    if (!filterText || nodeData[filterProps || 'id']?.toLowerCase().includes(lowerCaseFilter)) {
+      addNode(tree, folders, leafNode, { nodeData })
+    }
+  }
+
+  sortChildren(tree)
+
+  return postProcessing ? postProcessing(tree.childNodes!) : tree.childNodes!
+}
+
+export function traverseTree<T>(nodes: TreeNode<T>[], callback: (node: TreeNode<T>) => void) {
+  if (!nodes) {
+    return
+  }
+
+  for (const node of nodes) {
+    callback(node)
+    traverseTree(node.childNodes!, callback)
+  }
+}
+
+function addNode<T>(tree: TreeNode<T>, folders: TreeNode<T>[], leafNode: TreeNode<T>, data) {
+  for (const folderDesc of folders) {
+    let folder = _.find(tree.childNodes, x => x.id === folderDesc.id)
+    if (!folder) {
+      folder = { ...folderDesc, parent: tree, childNodes: [] }
+      tree.childNodes?.push(folder)
+    }
+    tree = folder
+  }
+
+  tree.childNodes?.push({ ...leafNode, parent: tree, ...data })
+}
+
+const sortChildren = tree => {
+  if (!tree.childNodes) {
+    return
+  }
+
+  tree.childNodes.sort((a, b) => {
+    if (a.type === b.type) {
+      return a.name < b.name ? -1 : 1
+    }
+    if (a.type === 'folder') {
+      return -1
+    } else {
+      return 1
+    }
+  })
+
+  tree.childNodes.forEach(sortChildren)
+}
+
+export const usePrevious = value => {
+  const ref = useRef()
+
+  useEffect(() => {
+    ref.current = value
+  }, [value])
+
+  return ref.current
+}
